requirements.txt
Django==2.2.4
django-environ==0.4.5
sqlparse==0.3.0
django-import-export==1.2.0
django-tables2==2.1.0
django-filter==2.2.0
django-crispy-forms==1.7.2
django-microsoft-auth==2.3.1
django-jet==1.0.8
django-constance==2.4.0
django-picklefield==2.0
simple-salesforce==0.74.3
djangorestframework==3.10.3
